requirements.txt
opencv-python
mediapipe
pyserial
pyttsx3
pillow